refactor: break up config models
