+++ src/codegen/cli/auth/session.py
@@ -3,10 +3,9 @@ from pathlib import Path
 from pygit2.repository import Repository
 
 from codegen.cli.git.repo import get_git_repo
-from codegen.shared.configs.config import load
 from codegen.shared.configs.constants import CODEGEN_DIR_NAME, CONFIG_FILENAME
-from codegen.shared.configs.global_config import config as global_config
-from codegen.shared.configs.models import Config
+from codegen.shared.configs.models.session import SessionConfig
+from codegen.shared.configs.session_configs import global_config, load_session_config
 
 
 class CodegenSession:
@@ -14,14 +13,14 @@ class CodegenSession:
 
     repo_path: Path  # TODO: rename to root_path
     codegen_dir: Path
-    config: Config
+    config: SessionConfig
     existing: bool
 
     def __init__(self, repo_path: Path):
         self.repo_path = repo_path
         self.codegen_dir = repo_path / CODEGEN_DIR_NAME
         self.existing = global_config.get_session(repo_path) is not None
-        self.config = load(self.codegen_dir / CONFIG_FILENAME)
+        self.config = load_session_config(self.codegen_dir / CONFIG_FILENAME)
         global_config.set_active_session(repo_path)
 
     @classmethod
