+++ src/codegen/shared/configs/models/global_session.py
@@ -1,6 +1,5 @@
 """Global config to manage different codegen sessions, as well as user auth."""
 
-# TODO: rename this file to global.py
 import json
 from pathlib import Path
 
@@ -39,18 +38,3 @@ class GlobalSessionConfig(BaseSettings):
 
         with open(SESSION_FILE, "w") as f:
             json.dump(self.model_dump(), f)
-
-
-def _load_global_config() -> GlobalSessionConfig:
-    """Load configuration from the JSON file."""
-    if SESSION_FILE.exists():
-        with open(SESSION_FILE) as f:
-            json_config = json.load(f)
-            return GlobalSessionConfig.model_validate(json_config, strict=False)
-
-    new_config = GlobalSessionConfig(sessions=[])
-    new_config.save()
-    return new_config
-
-
-config = _load_global_config()
